burningman/requirements.txt
selenium
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
lxml
python_anticaptcha
tinydb